annotation code refactor
